refactor: promote the cross-build handoff routing component to its own directory
Five files sharing the handoff-routing prefix each own a distinct facet of one
bounded responsibility, which is the §7 subdivision trigger. The directory is
also what makes membership explicit: handoff.ts is a different meaning of
"handoff" and stays outside, and repair-operation.ts joins because its
lifecycle classification decides whether a command receives a routing lifecycle
at all — the boundary question the deferral was waiting on.

Pure move: every moved file is byte-identical to its original modulo import
depth. Closes docs/todo/handoff-routing-subdivision.md.
