+++ docs/todo/README.md
@@ -255,7 +255,6 @@ guard stayed where it was.
 | ------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | [`unit-suite-concurrency-and-real-time-tests.md`](./unit-suite-concurrency-and-real-time-tests.md) | **Store half closed and removed; the real-time half is what is left.** The concurrent tier no longer opens a file-backed store — the doors decide from the opened handle and an invariant keeps them the only way in — so the load that justified converting real sleeps is smaller than when this was written. Its start condition is a flake observed under current conditions, naming which of the 22 remaining sites produced it. |
 | [`source-import-converter-cohesion.md`](./source-import-converter-cohesion.md) | **Startable now.** Five concerns at one layer in a 1058-line file; four converter classes are the documented subdivision trigger. A local fix improved its functions and grew the file — that is the datum.  |
-| [`handoff-routing-subdivision.md`](./handoff-routing-subdivision.md) | **Start after its boundary is settled.** This branch added the fourth cohesive cross-build routing facet, so §7's subdivision trigger is met. Promotion is deferred to a dedicated branch because `handoff-repair-operation.ts` now owns lifecycle-deciding classification as well as shared command grammar; that ownership decides the directory's shape. |
 | [`invariant-scans-stop-at-src.md`](./invariant-scans-stop-at-src.md)           | **Startable now.** One of two scans extended to `clients/hooks/` and found nothing; the other needs its detector taught a second idiom first. Measurement already done: three files, one alternate spelling. |
 
 ---
