+++ src/cli/handoff-notice.ts
@@ -1,4 +1,4 @@
-import type { HandoffPublicationIncident, HandoffSuccess } from '../coordinator/handoff-runner.js';
+import type { HandoffPublicationIncident, HandoffSuccess } from '../coordinator/handoff-routing/runner.js';
 import { formatHandoffPublicationIncident } from './format/handoff-publication.js';
 
 let noticeRendered = false;
